refactor(media): split media protocol internals
